+++ src/api/entities/src/orderProducts/orderProducts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaClient, Prisma } from '@prisma/client';
 
 @Injectable()
@@ -6,13 +6,21 @@ export class OrderProductsService {
     private prisma = new PrismaClient();
 
     async findAll(): Promise<any[]> {
-        return this.prisma.orderProduct.findMany();
+        const orderProducts = await this.prisma.orderProduct.findMany();
+        if (!orderProducts) {
+            throw new NotFoundException('OrderProducts not found');
+        }
+        return orderProducts;
     }
 
     async findOne(orderUuid: string, productUuid: string): Promise<any> {
-        return this.prisma.orderProduct.findUnique({
-            where: { order_uuid_product_uuid: { order_uuid: orderUuid, product_uuid: productUuid } },
+        const orderProduct = await this.prisma.orderProduct.findUnique({
+            where: { order_uuid_product_uuid: { order_uuid: orderUuid, product_uuid: productUuid } }
         });
+        if (!orderProduct) {
+            throw new NotFoundException('OrderProduct not found');
+        }
+        return orderProduct;
     }
 
     async create(createOrderProductDto: Prisma.OrderProductCreateInput): Promise<any> {
@@ -22,6 +30,12 @@ export class OrderProductsService {
     }
 
     async update(orderUuid: string, productUuid: string, updateOrderProductDto: Prisma.OrderProductUpdateInput): Promise<any> {
+        const existingOrderProduct = await this.prisma.orderProduct.findUnique({
+            where: { order_uuid_product_uuid: { order_uuid: orderUuid, product_uuid: productUuid } }
+        });
+        if (!existingOrderProduct) {
+            throw new NotFoundException(`OrderProduct not found`);
+        }
         return this.prisma.orderProduct.update({
             where: { order_uuid_product_uuid: { order_uuid: orderUuid, product_uuid: productUuid } },
             data: updateOrderProductDto
@@ -29,6 +43,12 @@ export class OrderProductsService {
     }
 
     async delete(orderUuid: string, productUuid: string): Promise<any> {
+        const existingOrderProduct = await this.prisma.orderProduct.findUnique({
+            where: { order_uuid_product_uuid: { order_uuid: orderUuid, product_uuid: productUuid } }
+        });
+        if (!existingOrderProduct) {
+            throw new NotFoundException(`OrderProduct not found`);
+        }
         return this.prisma.orderProduct.delete({
             where: { order_uuid_product_uuid: { order_uuid: orderUuid, product_uuid: productUuid } }
         });
