+++ src/api/entities/src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaClient, Prisma } from '@prisma/client';
 
 @Injectable()
@@ -6,13 +6,35 @@ export class ProductsService {
     private prisma = new PrismaClient();
 
     async findAll(): Promise<any[]> {
-        return this.prisma.product.findMany();
+        const products = await this.prisma.product.findMany();
+        if (!products) {
+            throw new NotFoundException('Products not found');
+        }
+        return products;
     }
 
     async findOne(uuid: string): Promise<any> {
-        return this.prisma.product.findUnique({
-            where: { uuid: uuid, },
+        const product = await this.prisma.product.findUnique({
+            where: { uuid: uuid },
         });
+        if (!product) {
+            throw new NotFoundException('Product not found');
+        }
+        return product;
+    }
+
+    async findOrders(uuid: string): Promise<any> {
+        const product = await this.prisma.product.findUnique({
+            where: { uuid: uuid },
+            select: { ProductOrders: {include: {Product: true,},}, },
+        });
+        if (!product) {
+            throw new NotFoundException('Product not found');
+        }
+        if (product.ProductOrders.length === 0) {
+            throw new NotFoundException('This product has no orders');
+        }
+        return product.ProductOrders;
     }
 
     async create(createProductDto: Prisma.ProductCreateInput): Promise<any> {
@@ -22,6 +44,12 @@ export class ProductsService {
     }
 
     async update(uuid: string, updateProductDto: Prisma.ProductUpdateInput): Promise<any> {
+        const existingProduct = await this.prisma.product.findUnique({
+            where: { uuid: uuid },
+        });
+        if (!existingProduct) {
+            throw new NotFoundException(`Product not found`);
+        }
         return this.prisma.product.update({
             where: { uuid: uuid },
             data: updateProductDto
@@ -29,6 +57,12 @@ export class ProductsService {
     }
 
     async delete(uuid: string): Promise<any> {
+        const existingProduct = await this.prisma.product.findUnique({
+            where: { uuid: uuid },
+        });
+        if (!existingProduct) {
+            throw new NotFoundException(`Product not found`);
+        }
         return this.prisma.product.delete({
             where: { uuid: uuid }
         });
